refactor: rename L2 Meta-Agent Bridge to L1 Relay Bridge with compat shims
Completes the code-level half of the #1065/#1236 terminology correction
per docs/architecture/LAYER_ARCHITECTURE.md's migration checklist: relay
agents are L1-resident (Orion Station) and only play the Layer 2 verifier
ROLE in the Triplex protocol — the "L2" naming was a canon error.

Key finding: src/bridges/l1_relay_bridge.py already existed (PR #416,
Nov 2025) but was a stale, never-imported fork — six months of later
work, including the mesh-runtime rewiring (#751), went into the old-named
file. This PR makes the mesh-backed implementation canonical under the
L1 name, replacing the stale fork, preserving live behavior exactly
(5 agents, mesh handshake sequence).

Backwards compatibility:
- src/bridges/l2_meta_agent_bridge.py and src/api/l2_meta_agent_api.py
  become deprecation shims re-exporting the same objects (l2_bridge IS
  l1_relay_bridge — one singleton).
- Routes served at /api/l1-relay-agents (canonical) AND /api/l2-agents
  (deprecated=True in OpenAPI), same handlers via a shared base router.
- Pre-existing test files pass unchanged through the shims except one
  mock target that follows the implementation.

Also: L1RelayAgent gains reality_layer="L1" / triplex_role=
"layer_2_verifier" (checklist item), inventory + router-coverage mapping
updated (new l1_relay_agents entry; l2_meta_agents marked deprecated
alias), API catalog snapshots regenerated (were stale), checklist boxes
ticked, and stray doc/code references updated.
