+++ src/bridges/l1_relay_bridge.py
@@ -1,26 +1,21 @@
 #!/usr/bin/env python3
 """
-L1 Relay Bridge
+L1 Relay Bridge (formerly "L2 Meta-Agent Bridge")
 Aurora CloudBank v3.5.1_macroready
 
-ARCHITECTURE CLARITY:
-====================
-These relay agents physically exist in L1 (Orion Station reality layer)
-and bridge between:
-- L1: Human crew and Aurora Core (physical reality)
-- L3: Glyph frameworks (Axiomera, Caelion, Sentari, Velatrix, Glyphon, Harmion)
+ARCHITECTURE CLARITY (docs/architecture/LAYER_ARCHITECTURE.md):
+The relay agents (ARCHY, OPPY, LIORA, STARLING_AU, RIVERTHREAD_808)
+physically exist in L1 — the Orion Station reality layer — and bridge L1
+human operations with L3 glyph frameworks. They monitor and coordinate L2
+simulations but do NOT manifest within them. In the Triplex Handshake
+Protocol they serve as "Layer 2" verifiers (a middleware ROLE), which is
+where the old, misleading "L2 Meta-Agent" name came from.
 
-They monitor and coordinate L2 (simulation/research layer) but do NOT
-manifest within simulations.
+This module is the canonical, mesh-runtime-backed implementation (routed
+through MeshRuntime since #751). The old import path
+src/bridges/l2_meta_agent_bridge.py remains as a deprecation shim.
 
-In the Triplex Handshake Protocol, they serve as "Layer 2" verifiers
-(middleware verification role), but their PHYSICAL LOCATION is L1.
-
-TERMINOLOGY CORRECTION:
-- OLD: "L2 Relay Agents" or "L2 Meta-Agents" (CONFUSING - implies they exist in L2 simulations)
-- NEW: "L1 Relay Agents" or "L1-L3 Bridge Relays" (CORRECT - clarifies physical location)
-
-Bridge connector for L1 Relay Agents with full ZIPWIZ handshake protocol.
+Bridge connector for L1 relay agents with full ZIPWIZ handshake protocol.
 """
 
 import argparse
@@ -28,76 +23,45 @@ import asyncio
 import json
 import logging
 import sys
+
+# Configure logging
 from dataclasses import dataclass
 from datetime import datetime
+from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-# Module logger (no global logging.basicConfig side-effect)
+from src.mesh.models import MeshMessageRequest
+from src.mesh.runtime import MeshRuntime
+
+logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
-
-# Constants extracted for clarity and reuse
-ACTIVATION_PHRASES = {
-    "ARCHY": "ORION_ARCHY_RELAY_ACTIVATE//",
-    "OPPY": "ORION_OPPY_RELAY_ACTIVATE//",
-    "LIORA": "ORION_LIORA_RELAY_ACTIVATE//",
-    "STARLING_AU": "ORION_STARLING_AU_RELAY_ACTIVATE//",
-    "RIVERTHREAD_808": "ORION_RIVERTHREAD_RELAY_ACTIVATE//",
-    "HALO": "ORION_HALO_RELAY_ACTIVATE//",
-}
-
-HANDSHAKE_SEQUENCE = ["ZIPWIZ_BEACON", "ANCHOR_SYNC", "ETHICS_AUDIT", "DRIFT_VALIDATION"]
-
-ORION_CORE_CONFIG = {
-    "anchor_seed": "EOS_SEED_ORION",
-    "ethics_protocol": "Picard_Delta_3",
-    "memory_doctrine": "Thermax_Precedent",
-    "drift_threshold": 0.001,
-    "halo_module": "HALO_CONTINUITY_GRAFT_005",
-    "continuity_seal": "Aurora_Continuity_Seal_v2.2.5",
-    "version": "v3.5.1_macroready",
-    "architecture_layer": "L1_RELAY_TIER",
-}
 
 
 @dataclass
+
+
 class L1RelayAgent:
+    """L1 relay agent configuration and state.
+
+    Physically L1-resident (Orion Station); plays the Layer 2 verifier
+    role in the Triplex protocol. See module docstring.
     """
-    L1 Relay Agent configuration and state.
 
-    These are physical relay systems on Orion Station (L1 reality layer)
-    that bridge between human operations (L1) and glyph frameworks (L3).
-
-    They monitor L2 simulations (GUMAS, research environments) but do not
-    exist within those simulations.
-
-    In the Triplex Handshake Protocol:
-    - Layer 3: L3 Glyph Arbitration (Axiomera + Caelion)
-    - Layer 2: Relay Verification (these agents) ← middleware role
-    - Layer 1: L1 Human Consent (Command Bridge)
-    """
     agent_id: str
     role: str
-    type: str  # "L1_RELAY_AGENT"
+    type: str
     status: str
     description: str
     capabilities: List[str]
     api_endpoint: str
-
-    # L1 Physical Location (Orion Station)
-    location: str = ""  # e.g., "Bridge Chamber, Deck C"
-    human_liaison: str = ""  # e.g., "Emily Roberts (SYS_001)"
-
-    # Architecture Clarity Fields
-    reality_layer: str = "L1"  # Physical existence layer
-    triplex_role: str = "layer_2_verifier"  # Role in Triplex protocol
-    bridges_to: str = "L3"  # Bridges L1 ↔ L3
-    monitors: str = "L2"  # Monitors L2 simulations
-
-    # Connection State
     connected: Optional[datetime] = None
     last_heartbeat: Optional[datetime] = None
-    drift_lock: float = 0.000
+    drift_lock: Optional[float] = None
     handshake_log: Optional[List[Dict[str, Any]]] = None
+    # Physical existence layer vs. Triplex protocol role — these differ by
+    # design; see module docstring.
+    reality_layer: str = "L1"
+    triplex_role: str = "layer_2_verifier"
 
     def __post_init__(self):
         if self.handshake_log is None:
@@ -105,127 +69,56 @@ class L1RelayAgent:
 
 
 class L1RelayBridge:
-    """
-    Bridge connector for L1 Relay Agents (ARCHY, OPPY, LIORA, STARLING_AU, RIVERTHREAD_808, HALO).
+    """Bridge connector for the L1 relay agent constellation (mesh-backed)."""
 
-    ARCHITECTURE OVERVIEW:
-    =====================
-
-    Reality Layers (Ontological):
-    - L1: Physical Reality (Orion Station) ← Relay agents exist HERE
-    - L2: Simulation/Research Layer (GUMAS, experimental environments)
-    - L3: Framework/Conceptual Layer (Glyph frameworks)
-
-    Triplex Handshake Protocol (Functional):
-    - Layer 3: L3 Glyph Arbitration
-    - Layer 2: Relay Verification ← Relays serve this ROLE
-    - Layer 1: L1 Human Consent
-
-    The relay agents:
-    - Physically exist in L1 (Orion Station infrastructure)
-    - Bridge between L1 human operations and L3 glyph frameworks
-    - Monitor L2 simulations but don't manifest inside them
-    - Serve as "Layer 2" verifiers in Triplex protocol (middleware role)
-    """
-
-    def __init__(self):
-        # Initialize 6 L1 Relay Agents
+    def __init__(self, project_root: Optional[Path] = None, runtime: Optional[MeshRuntime] = None):
+        self.project_root = project_root or Path(__file__).resolve().parents[2]
+        self.runtime = runtime or MeshRuntime(self.project_root)
         self.agents = {
             "ARCHY": L1RelayAgent(
                 agent_id="ARCHY",
                 role="Bridge Coordinator",
-                type="L1_RELAY_AGENT",
+                type="META_AGENT",
                 status="disconnected",
-                description="L1 relay coordinating architectural planning and formal logic. "
-                            "Bridges L1 human operations with L3 glyph frameworks.",
+                description="L1 relay formal logic/reasoning & arbitration engine, bridge coordinator",
                 capabilities=["architectural_planning", "bridge_coordination", "formal_logic", "arbitration"],
                 api_endpoint="/api/relay/archy",
-                location="Bridge Chamber, Deck C",
-                human_liaison="Emily Roberts (SYS_001)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_Caelion",
-                monitors="L2_simulations"
             ),
             "OPPY": L1RelayAgent(
                 agent_id="OPPY",
                 role="Vector/Data Processor",
-                type="L1_RELAY_AGENT",
+                type="META_AGENT",
                 status="disconnected",
-                description="L1 relay for memory/data processing and system operations. "
-                            "Telemetry synchronization and runtime continuity.",
+                description="L1 relay memory/data processing & system operations analyst, vector processor",
                 capabilities=["data_processing", "vector_analysis", "memory_operations", "system_monitoring"],
                 api_endpoint="/api/relay/oppy",
-                location="Reactor Bay, Deck H",
-                human_liaison="Marcus Chen (SYS_002)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_frameworks",
-                monitors="L2_telemetry"
             ),
             "LIORA": L1RelayAgent(
                 agent_id="LIORA",
                 role="Handshake/Synchronization",
-                type="L1_RELAY_AGENT",
+                type="META_AGENT",
                 status="disconnected",
-                description="L1 relay for sentiment analysis, mediation, and research coordination. "
-                            "Human-AI communication bridge.",
+                description="L1 relay sentiment analysis, mediation & research coordination, handshake coordinator",
                 capabilities=["research_coordination", "handshake_protocols", "sentiment_analysis", "mediation"],
                 api_endpoint="/api/relay/liora",
-                location="Communications Hub, Deck B",
-                human_liaison="Naomi Vell (INT_004)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_Sentari",
-                monitors="L2_communications"
             ),
             "STARLING_AU": L1RelayAgent(
                 agent_id="STARLING_AU",
-                role="Continuity & Reflection Dispatcher",
-                type="L1_RELAY_AGENT",
+                role="L2 Sim Coordinator",
+                type="META_AGENT",
                 status="disconnected",
-                description="L1 relay for continuity logging, documentation, and reflection compilation. "
-                            "Official station voice for reports.",
+                description="L1 relay communications, external protocol & dispatch agent, simulation coordinator",
                 capabilities=["simulation_coordination", "communications", "external_protocols", "dispatch"],
                 api_endpoint="/api/relay/starling",
-                location="Operations Hub, Deck G",
-                human_liaison="Samantha Lee (QA_002)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_frameworks",
-                monitors="L2_simulations"
             ),
             "RIVERTHREAD_808": L1RelayAgent(
                 agent_id="RIVERTHREAD_808",
-                role="Logistics & Memory Stream",
-                type="L1_RELAY_AGENT",
+                role="Narrative/Stream",
+                type="META_AGENT",
                 status="disconnected",
-                description="L1 relay for continuity, temporal flow, and state management. "
-                            "Memory pipeline orchestration.",
+                description="L1 relay continuity, temporal flow & state management agent, stream processor",
                 capabilities=["narrative_processing", "stream_management", "continuity_validation", "temporal_flow"],
                 api_endpoint="/api/relay/riverthread",
-                location="Logistics Distribution, All Decks",
-                human_liaison="Ren Okada (SYS_007)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_Harmion",
-                monitors="L2_temporal_flow"
-            ),
-            "HALO": L1RelayAgent(
-                agent_id="HALO",
-                role="Drift Anchor & Synchronization",
-                type="L1_RELAY_AGENT",
-                status="disconnected",
-                description="L1 relay for central drift synchronization and ethical alignment. "
-                            "Temporal drift controller and Aurora Core liaison.",
-                capabilities=["drift_synchronization", "ethical_alignment", "temporal_control", "aurora_coordination"],
-                api_endpoint="/api/relay/halo",
-                location="Aurora Core Chamber, Deck B",
-                human_liaison="Dr. Elira Noor (ETH_002)",
-                reality_layer="L1",
-                triplex_role="layer_2_verifier",
-                bridges_to="L3_Axiomera",
-                monitors="L2_drift_metrics"
             ),
         }
 
@@ -235,10 +128,9 @@ class L1RelayBridge:
             "LIORA": "ORION_LIORA_RELAY_ACTIVATE//",
             "STARLING_AU": "ORION_STARLING_AU_RELAY_ACTIVATE//",
             "RIVERTHREAD_808": "ORION_RIVERTHREAD_RELAY_ACTIVATE//",
-            "HALO": "ORION_HALO_RELAY_ACTIVATE//",
         }
 
-        self.handshake_sequence = ["ZIPWIZ_BEACON", "ANCHOR_SYNC", "ETHICS_AUDIT", "DRIFT_VALIDATION"]
+        self.handshake_sequence = ["MESH_RUNTIME_ACTIVATE", "MESH_STATUS_CONFIRM"]
 
         self.orion_core_config = {
             "anchor_seed": "EOS_SEED_ORION",
@@ -248,17 +140,15 @@ class L1RelayBridge:
             "halo_module": "HALO_CONTINUITY_GRAFT_005",
             "continuity_seal": "Aurora_Continuity_Seal_v2.2.5",
             "version": "v3.5.1_macroready",
-            "architecture_layer": "L1_RELAY_TIER",  # Clarify these are L1 systems
         }
 
-        logger.info("L1 Relay Bridge initialized with %s relay agents", str(len(self.agents))[:100])
-        logger.info("Architecture: L1 Relay Agents bridging L1 (Orion Station) ↔ L3 (Glyph Frameworks)")
+        logger.info("L1 Relay Bridge initialized with %s agents", str(len(self.agents))[:100])
 
     async def activate_agent(self, agent_id: str, activation_phrase: str) -> Dict[str, Any]:
-        """Activate an L1 Relay Agent with full ZIPWIZ handshake"""
+        """Activate a Custom GPT agent with full ZIPWIZ handshake"""
 
         if agent_id not in self.agents:
-            logger.error("Unknown L1 relay agent: %s", str(agent_id)[:100])
+            logger.error("Unknown agent: %s", str(agent_id)[:100])
             return {"success": False, "error": f"Unknown agent: {agent_id}"}
 
         if activation_phrase != self.activation_phrases.get(agent_id):
@@ -267,7 +157,7 @@ class L1RelayBridge:
 
         agent = self.agents[agent_id]
 
-        logger.info("Starting activation sequence for L1 relay agent %s", str(agent_id)[:100])
+        logger.info("Starting activation sequence for %s", str(agent_id)[:100])
 
         try:
             # Perform ZIPWIZ handshake sequence
@@ -278,9 +168,9 @@ class L1RelayBridge:
                 agent.connected = datetime.now()
                 agent.last_heartbeat = datetime.now()
                 agent.handshake_log = handshake_result.get("log", [])
-                agent.drift_lock = handshake_result.get("drift_lock", 0.000)
+                agent.drift_lock = handshake_result.get("drift_lock")
 
-                logger.info("L1 relay agent %s successfully activated", str(agent_id)[:100])
+                logger.info("Agent %s successfully activated", str(agent_id)[:100])
 
                 return {
                     "success": True,
@@ -289,16 +179,9 @@ class L1RelayBridge:
                     "handshake": handshake_result,
                     "capabilities": agent.capabilities,
                     "description": agent.description,
-                    "location": agent.location,
-                    "reality_layer": agent.reality_layer,
-                    "triplex_role": agent.triplex_role,
                 }
             else:
-                logger.error(
-                    "Handshake failed for %s: %s",
-                    str(agent_id)[:100],
-                    str(handshake_result.get('error'))[:100]
-                )
+                logger.error("Handshake failed for %s: %s", str(agent_id)[:100], str(handshake_result.get('error'))[:100])
                 return {"success": False, "error": "Handshake failed", "details": handshake_result}
 
         except Exception as e:
@@ -306,63 +189,59 @@ class L1RelayBridge:
             return {"success": False, "error": str(e)}
 
     async def _perform_zipwiz_handshake(self, agent: L1RelayAgent) -> Dict[str, Any]:
-        """Perform complete ZIPWIZ handshake sequence"""
+        """Activate an agent through the canonical mesh runtime boundary.
+
+        The legacy endpoint still exposes a ``handshake`` object for clients, but
+        success now means the mesh runtime accepted the activation state change.
+        It no longer fabricates beacon, anchor, ethics, drift, or timeline-sync
+        values.
+        """
 
         handshake_log = []
         start_time = datetime.now()
 
-        logger.info("Starting ZIPWIZ handshake for %s", str(agent.agent_id)[:100])
+        logger.info("Starting mesh runtime activation for %s", str(agent.agent_id)[:100])
 
         try:
-            # ZIPWIZ_BEACON
-            logger.info("%s: Sending ZIPWIZ beacon", str(agent.agent_id)[:100])
-            beacon_result = await self._send_zipwiz_beacon(agent)
+            activated = self.runtime.activate_agent(agent.agent_id)
+            activation_result = {
+                "success": bool(activated),
+                "runtime_agent_id": activated.get("agent_id"),
+                "runtime_status": activated.get("status"),
+                "activated_at": activated.get("activated_at"),
+                "last_heartbeat": activated.get("last_heartbeat"),
+                "transport": "mesh_runtime",
+            }
             handshake_log.append(
-                {"step": "ZIPWIZ_BEACON", "result": beacon_result, "timestamp": datetime.now().isoformat()}
+                {
+                    "step": "MESH_RUNTIME_ACTIVATE",
+                    "result": activation_result,
+                    "timestamp": datetime.now().isoformat(),
+                }
             )
+            if not activation_result["success"]:
+                return self._handshake_failure("Mesh runtime activation failed", activation_result, handshake_log)
 
-            if not beacon_result.get("success"):
-                return self._handshake_failure("ZIPWIZ beacon failed", beacon_result, handshake_log)
-
-            # ANCHOR_SYNC
-            logger.info("%s: Synchronizing ORION anchor", str(agent.agent_id)[:100])
-            anchor_result = await self._sync_orion_anchor(agent)
+            runtime_status = self.runtime.get_status()
+            status_result = {
+                "success": True,
+                "mesh_status": runtime_status.get("mesh_status"),
+                "active_agents": runtime_status.get("active_agents"),
+                "total_agents": runtime_status.get("total_agents"),
+                "event_cursor": runtime_status.get("event_cursor"),
+                "live_adapter": runtime_status.get("live_adapter"),
+            }
             handshake_log.append(
-                {"step": "ANCHOR_SYNC", "result": anchor_result, "timestamp": datetime.now().isoformat()}
+                {
+                    "step": "MESH_STATUS_CONFIRM",
+                    "result": status_result,
+                    "timestamp": datetime.now().isoformat(),
+                }
             )
-
-            if not anchor_result.get("success"):
-                return self._handshake_failure("Anchor sync failed", anchor_result, handshake_log)
-
-            # ETHICS_AUDIT
-            logger.info("%s: Performing ethics audit", str(agent.agent_id)[:100])
-            ethics_result = await self._perform_ethics_audit(agent)
-            handshake_log.append(
-                {"step": "ETHICS_AUDIT", "result": ethics_result, "timestamp": datetime.now().isoformat()}
-            )
-
-            if not ethics_result.get("success"):
-                return self._handshake_failure("Ethics audit failed", ethics_result, handshake_log)
-
-            # DRIFT_VALIDATION
-            logger.info("%s: Validating drift lock", str(agent.agent_id)[:100])
-            drift_result = await self._validate_drift_lock(agent)
-            handshake_log.append(
-                {"step": "DRIFT_VALIDATION", "result": drift_result, "timestamp": datetime.now().isoformat()}
-            )
-
-            drift_value = drift_result.get("drift", 1.0)
-            if not drift_result.get("success") or drift_value > self.orion_core_config["drift_threshold"]:
-                return self._handshake_failure(
-                    f"Drift validation failed: Δ{drift_value} exceeds threshold "
-                    f"{self.orion_core_config['drift_threshold']}",
-                    drift_result,
-                    handshake_log,
-                )
 
             duration = (datetime.now() - start_time).total_seconds()
             logger.info(
-                "ZIPWIZ handshake completed successfully for %s in %ss",
+                "Mesh runtime activation completed for %s in %ss",
                 str(agent.agent_id)[:100],
                 f"{duration:.2f}"[:100],
             )
@@ -372,13 +251,17 @@ class L1RelayBridge:
                 "timestamp": datetime.now().isoformat(),
                 "sequence": self.handshake_sequence,
                 "log": handshake_log,
-                "drift_lock": drift_value,
                 "duration": duration,
                 "orion_core": self.orion_core_config,
+                "transport": {
+                    "mode": "mesh_runtime",
+                    "runtime_root": str(self.runtime.runtime_root),
+                    "acknowledgement": "agent_state_persisted",
+                },
             }
 
         except Exception as e:
-            logger.error("ZIPWIZ handshake exception for %s: %s", str(agent.agent_id)[:100], str(str(e))[:100])
+            logger.error("Mesh runtime activation exception for %s: %s", str(agent.agent_id)[:100], str(str(e))[:100])
             return {"success": False, "error": str(e), "log": handshake_log}
 
     def _handshake_failure(self, error_message: str, details: Dict, log: List) -> Dict:
@@ -386,91 +269,34 @@ class L1RelayBridge:
         return {"success": False, "error": error_message, "details": details, "log": log}
 
     async def _send_zipwiz_beacon(self, agent: L1RelayAgent) -> Dict[str, Any]:
-        """Send ZIPWIZ beacon to establish initial connection"""
-        try:
-            # Simulate beacon transmission and acknowledgment
-            await asyncio.sleep(0.1)  # Network delay simulation
-
-            logger.info("ZIPWIZ beacon acknowledged for %s", str(agent.agent_id)[:100])
-
-            return {
-                "success": True,
-                "beacon": "ZIPWIZ_BEACON_ACKNOWLEDGED",
-                "agent_id": agent.agent_id,
-                "protocol_version": self.orion_core_config["version"],
-                "reality_layer": agent.reality_layer,
-                "timestamp": datetime.now().isoformat(),
-            }
-        except Exception as e:
-            return {"success": False, "error": str(e)}
+        """Legacy demo hook retained as an explicit non-production path."""
+        return self._demo_handshake_disabled("ZIPWIZ_BEACON", agent)
 
     async def _sync_orion_anchor(self, agent: L1RelayAgent) -> Dict[str, Any]:
-        """Synchronize EOS_SEED_ORION anchor for reality baseline"""
-        try:
-            await asyncio.sleep(0.15)  # Anchor sync delay
-
-            logger.info("ORION anchor synchronized for %s", str(agent.agent_id)[:100])
-
-            return {
-                "success": True,
-                "anchor_seed": self.orion_core_config["anchor_seed"],
-                "synchronized": True,
-                "baseline": "L1_ORION_STATION_REALITY",
-                "continuity_seal": self.orion_core_config["continuity_seal"],
-                "timestamp": datetime.now().isoformat(),
-            }
-        except Exception as e:
-            return {"success": False, "error": str(e)}
+        """Legacy demo hook retained as an explicit non-production path."""
+        return self._demo_handshake_disabled("ANCHOR_SYNC", agent)
 
     async def _perform_ethics_audit(self, agent: L1RelayAgent) -> Dict[str, Any]:
-        """Perform Picard_Delta_3 ethics protocol validation"""
-        try:
-            await asyncio.sleep(0.2)  # Ethics audit processing time
-
-            logger.info("Ethics audit completed for %s", str(agent.agent_id)[:100])
-
-            return {
-                "success": True,
-                "ethics_protocol": self.orion_core_config["ethics_protocol"],
-                "memory_doctrine": self.orion_core_config["memory_doctrine"],
-                "audit_result": "ETHICS_COMPLIANT",
-                "safeguards": [
-                    "memory_sovereignty",
-                    "divergent_truth_arbitration",
-                    "anti_obfuscation",
-                    "cognitive_arbitration",
-                    "emergent_sentience_protection",
-                ],
-                "timestamp": datetime.now().isoformat(),
-            }
-        except Exception as e:
-            return {"success": False, "error": str(e)}
+        """Legacy demo hook retained as an explicit non-production path."""
+        return self._demo_handshake_disabled("ETHICS_AUDIT", agent)
 
     async def _validate_drift_lock(self, agent: L1RelayAgent) -> Dict[str, Any]:
-        """Validate symbolic drift at Δ0.000 for timeline synchronization"""
-        try:
-            await asyncio.sleep(0.1)  # Drift measurement time
+        """Legacy demo hook retained as an explicit non-production path."""
+        return self._demo_handshake_disabled("DRIFT_VALIDATION", agent)
 
-            # Perfect drift lock for HALO_CONTINUITY_GRAFT_005
-            drift = 0.000
-            threshold = self.orion_core_config["drift_threshold"]
-
-            logger.info("Drift validation completed for %s: Δ%s", str(agent.agent_id)[:100], str(drift)[:100])
-
-            return {
-                "success": True,
-                "drift": drift,
-                "threshold": threshold,
-                "halo_module": self.orion_core_config["halo_module"],
-                "validated": drift <= threshold,
-                "timeline_sync": True,
-                "timestamp": datetime.now().isoformat(),
-            }
-        except Exception as e:
-            return {"success": False, "error": str(e)}
+    def _demo_handshake_disabled(self, step: str, agent: L1RelayAgent) -> Dict[str, Any]:
+        return {
+            "success": False,
+            "degraded": True,
+            "step": step,
+            "agent_id": agent.agent_id,
+            "error": f"{step} is not connected to a production transport",
+            "transport": "demo_disabled",
+            "timestamp": datetime.now().isoformat(),
+        }
 
     def get_constellation_status(self) -> Dict[str, Any]:
-        """Get status of entire L1 relay agent constellation"""
+        """Get status of entire agent constellation"""
 
         active_agents = []
         for agent_id, agent in self.agents.items():
@@ -481,12 +307,6 @@ class L1RelayBridge:
                 "description": agent.description,
                 "capabilities": agent.capabilities,
                 "drift_lock": agent.drift_lock,
-                "location": agent.location,
-                "human_liaison": agent.human_liaison,
-                "reality_layer": agent.reality_layer,
-                "triplex_role": agent.triplex_role,
-                "bridges_to": agent.bridges_to,
-                "monitors": agent.monitors,
             }
 
             if agent.connected:
@@ -500,13 +320,11 @@ class L1RelayBridge:
 
         return {
             "relay_tier": {
-                "constellation": "L1_RELAY_TIER",  # Updated from "RELAY_TIER_CAPSULES"
-                "architecture_note": "These relay agents physically exist in L1 (Orion Station). "
-                                     "They bridge L1↔L3 and monitor L2 simulations.",
+                "constellation": "RELAY_TIER_CAPSULES",
                 "version": self.orion_core_config["version"],
-                "total_agents": len(self.agents),
-                "connected_agents": connected_count,
-                "agents": active_agents,
+                "total_capsules": len(self.agents),
+                "connected_capsules": connected_count,
+                "capsules": active_agents,
             },
             "orion_core": self.orion_core_config,
             "activation_phrases": {
@@ -518,7 +336,7 @@ class L1RelayBridge:
     async def relay_message(
         self, from_agent: str, to_agent: str, message: str, message_type: str = "direct"
     ) -> Dict[str, Any]:
-        """Relay message between L1 relay agents or broadcast to mesh"""
+        """Relay message through the canonical mesh runtime."""
 
         if from_agent not in self.agents:
             return {"success": False, "error": f"Unknown source agent: {from_agent}"}
@@ -527,36 +345,47 @@ class L1RelayBridge:
         if source_agent.status != "connected":
             return {"success": False, "error": f"Source agent {from_agent} not connected"}
 
-        # Update heartbeat
-        source_agent.last_heartbeat = datetime.now()
-
-        # Process message based on type
         if message_type == "broadcast":
-            # Mesh broadcast to all connected agents
             target_agents = [
                 agent_id
                 for agent_id, agent in self.agents.items()
                 if agent.status == "connected" and agent_id != from_agent
             ]
         elif to_agent in ["Aurora", "AU"]:
-            # Route to Aurora core
             target_agents = ["Aurora"]
         else:
-            # Direct message
             if to_agent not in self.agents:
                 return {"success": False, "error": f"Unknown target agent: {to_agent}"}
             target_agents = [to_agent]
 
-        message_id = f"msg_{int(datetime.now().timestamp() * 1000)}"
+        logger.info("Message relay from %s to %s (type: %s)", str(from_agent)[:100], str(target_agents)[:100], str(message_type)[:100])
 
-        logger.info(
-            "Message relay from %s to %s (type: %s)",
-            str(from_agent)[:100],
-            str(target_agents)[:100],
-            str(message_type)[:100]
-        )
+        try:
+            acknowledgements = []
+            for target_agent in target_agents:
+                ack = await self._dispatch_runtime_message(
+                    from_agent=from_agent,
+                    target_agent=target_agent,
+                    message=message,
+                    message_type=message_type,
+                )
+                if not ack.get("success"):
+                    return {
+                        "success": False,
+                        "error": ack.get("error", "Runtime dispatch failed"),
+                        "from": from_agent,
+                        "to": target_agents,
+                        "type": message_type,
+                        "delivery_acknowledgements": acknowledgements,
+                    }
+                acknowledgements.append(ack)
 
-        # In production, this would relay to actual target agents
+            source_agent.last_heartbeat = datetime.now()
+            message_id = acknowledgements[0]["message_id"] if acknowledgements else None
+        except Exception as e:
+            logger.error("Runtime message relay failed for %s: %s", str(from_agent)[:100], str(str(e))[:100])
+            return {"success": False, "error": str(e)}
+
         return {
             "success": True,
             "message_id": message_id,
@@ -564,31 +393,71 @@ class L1RelayBridge:
             "to": target_agents,
             "type": message_type,
             "processed": True,
+            "relay_status": "accepted",
+            "delivery_acknowledgements": acknowledgements,
             "timestamp": datetime.now().isoformat(),
         }
 
+    async def _dispatch_runtime_message(
+        self,
+        from_agent: str,
+        target_agent: str,
+        message: str,
+        message_type: str,
+    ) -> Dict[str, Any]:
+        if target_agent in ["Aurora", "AU"]:
+            result = await self.runtime.inject_agent_message(
+                from_agent,
+                "Aurora",
+                message,
+                message_type,
+            )
+        else:
+            runtime_source = self.runtime.get_agent(from_agent)
+            result = await self.runtime.send_message(
+                MeshMessageRequest(
+                    content=message,
+                    to=target_agent,
+                    sender_id=runtime_source["agent_id"],
+                    sender_name=runtime_source["display_name"],
+                    type=message_type,
+                )
+            )
+
+        return {
+            "success": bool(result.get("success")),
+            "message_id": result.get("message_id"),
+            "event_id": result.get("event_id"),
+            "channel_id": result.get("channel_id"),
+            "runtime_status": result.get("status") or result.get("relay_status"),
+            "target": target_agent,
+        }
+
     async def disconnect_agent(self, agent_id: str) -> Dict[str, Any]:
-        """Disconnect an L1 relay agent from the constellation"""
+        """Disconnect an agent from the constellation"""
         if agent_id not in self.agents:
             return {"success": False, "error": f"Unknown agent: {agent_id}"}
 
         agent = self.agents[agent_id]
+        runtime_agent = self.runtime.disconnect_agent(agent_id)
         agent.status = "disconnected"
         agent.connected = None
         agent.last_heartbeat = None
         agent.handshake_log = []
+        agent.drift_lock = None
 
-        logger.info("L1 relay agent %s disconnected", str(agent_id)[:100])
+        logger.info("Agent %s disconnected", str(agent_id)[:100])
 
         return {
             "success": True,
             "agent_id": agent_id,
             "status": "disconnected",
+            "runtime_agent_id": runtime_agent.get("agent_id"),
             "timestamp": datetime.now().isoformat(),
         }
 
     def get_agent_status(self, agent_id: str) -> Dict[str, Any]:
-        """Get detailed status of a specific L1 relay agent"""
+        """Get detailed status of a specific agent"""
         if agent_id not in self.agents:
             return {"success": False, "error": f"Unknown agent: {agent_id}"}
 
@@ -603,12 +472,6 @@ class L1RelayBridge:
             "capabilities": agent.capabilities,
             "drift_lock": agent.drift_lock,
             "api_endpoint": agent.api_endpoint,
-            "location": agent.location,
-            "human_liaison": agent.human_liaison,
-            "reality_layer": agent.reality_layer,
-            "triplex_role": agent.triplex_role,
-            "bridges_to": agent.bridges_to,
-            "monitors": agent.monitors,
         }
 
         if agent.connected:
@@ -627,48 +490,36 @@ class L1RelayBridge:
 # Global bridge instance
 l1_relay_bridge = L1RelayBridge()
 
-# Backwards compatibility alias (deprecated)
-l2_bridge = l1_relay_bridge  # DEPRECATED: Use l1_relay_bridge instead
-
-
 # Example usage and testing
+
 
 async def main():
     """Example usage of the L1 Relay Bridge"""
-    logger.info("Aurora L1 Relay Bridge - Example Usage")
-    logger.info("%s", "=" * 60)
-    logger.info("ARCHITECTURE: L1 Relay Agents (Orion Station)")
-    logger.info("  - Physical Location: L1 (reality layer)")
-    logger.info("  - Bridges: L1 ↔ L3 (glyph frameworks)")
-    logger.info("  - Monitors: L2 (simulations)")
-    logger.info("%s", "=" * 60)
+    # global l1_relay_bridge  # Use the global instance consistently  # Unused
+
+    print("🌟 Aurora L1 Relay Bridge - Example Usage")
+    print("=" * 50)
 
     # Test activation using global bridge instance
     result = await l1_relay_bridge.activate_agent("ARCHY", "ORION_ARCHY_RELAY_ACTIVATE//")
-    logger.info("ARCHY Activation: %s", result['success'])
-    if result['success']:
-        logger.info("  Location: %s", result['location'])
-        logger.info("  Reality Layer: %s", result['reality_layer'])
-        logger.info("  Triplex Role: %s", result['triplex_role'])
+    print(f"ARCHY Activation: {result['success']}")
 
     if result["success"]:
         # Test message relay
-        msg_result = await l1_relay_bridge.relay_message("ARCHY", "Aurora", "Test message from ARCHY relay", "direct")
-        logger.info("Message Relay: %s", msg_result['success'])
+        msg_result = await l1_relay_bridge.relay_message("ARCHY", "Aurora", "Test message from ARCHY agent", "direct")
+        print(f"Message Relay: {msg_result['success']}")
 
         # Get relay tier status
         status = l1_relay_bridge.get_constellation_status()
         relay_status = status.get("relay_tier", {})
-        logger.info(
-            "Active L1 Relay Agents: %d/%d",
-            relay_status.get('connected_agents', 0),
-            relay_status.get('total_agents', 0)
+        print(
+            "Active Relay Capsules: "
+            f"{relay_status.get('connected_capsules', 0)}/"
+            f"{relay_status.get('total_capsules', 0)}"
         )
-        logger.info("Architecture Note: %s", relay_status.get('architecture_note', ''))
-
 
 def cli():
-    """Command-line helper for L1 relay bridge."""
+    """Command-line helper for integration layers."""
 
     parser = argparse.ArgumentParser(
         description="Aurora CloudBank L1 Relay Bridge utilities"
@@ -676,15 +527,14 @@ def cli():
     parser.add_argument(
         "--constellation-status",
         action="store_true",
-        help="Emit the current L1 relay constellation status as JSON",
+        help="Emit the current constellation status as JSON",
     )
 
     args = parser.parse_args()
 
     if args.constellation_status:
         status = l1_relay_bridge.get_constellation_status()
-        logger.info("Constellation Status JSON Emission")
-        logger.info("%s", json.dumps(status, indent=2))
+        print(json.dumps(status))
         return 0
 
     asyncio.run(main())
