+++ tests/test_api_router_coverage.py
@@ -69,7 +69,8 @@ ROUTER_SYMBOL_TO_INVENTORY_ID = {
     "r2_telemetry_router": "r2_telemetry",
     "crew_agents_router": "crew_agents",
     "l1_station_router": "l1_station",
-    "l2_meta_agent_router": "l2_meta_agents",
+    "l1_relay_router": "l1_relay_agents",
+    "l1_relay_legacy_router": "l2_meta_agents",
     "drift_metrics_router": "drift_metrics",
     "build_sensor_router(_sensor_array)": "sensor_array",
     "qgia_router": "qgia_forecast",
